refactor(analysis-types): extract analysis receipt contract
