+++ crates/tokmd-analysis-types/src/lib.rs
@@ -32,12 +32,10 @@ mod fun;
 mod git;
 mod imports;
 mod license;
+mod receipt;
 mod source;
 mod topics;
 pub mod util;
-
-use serde::{Deserialize, Serialize};
-use tokmd_types::{ScanStatus, ToolInfo};
 
 pub use api_surface::{ApiExportItem, ApiSurfaceReport, LangApiSurface, ModuleApiRow};
 pub use archetype::Archetype;
@@ -80,6 +78,7 @@ pub use git::{
 };
 pub use imports::{ImportEdge, ImportReport};
 pub use license::{LicenseFinding, LicenseReport, LicenseSourceKind};
+pub use receipt::AnalysisReceipt;
 pub use source::AnalysisSource;
 pub use topics::{TopicClouds, TopicTerm};
 pub use util::{
@@ -95,34 +94,6 @@ pub use tokmd_scan::{gini_coefficient, percentile, round_f64, safe_ratio};
 /// v8: Near-dup clusters, selection metadata, max_pairs guardrail, runtime stats.
 /// v9: Added effort estimation report.
 pub const ANALYSIS_SCHEMA_VERSION: u32 = 9;
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AnalysisReceipt {
-    pub schema_version: u32,
-    pub generated_at_ms: u128,
-    pub tool: ToolInfo,
-    pub mode: String,
-    pub status: ScanStatus,
-    pub warnings: Vec<String>,
-    pub source: AnalysisSource,
-    pub args: AnalysisArgsMeta,
-    pub archetype: Option<Archetype>,
-    pub topics: Option<TopicClouds>,
-    pub entropy: Option<EntropyReport>,
-    pub predictive_churn: Option<PredictiveChurnReport>,
-    pub corporate_fingerprint: Option<CorporateFingerprint>,
-    pub license: Option<LicenseReport>,
-    pub derived: Option<DerivedReport>,
-    pub assets: Option<AssetReport>,
-    pub deps: Option<DependencyReport>,
-    pub git: Option<GitReport>,
-    pub imports: Option<ImportReport>,
-    pub dup: Option<DuplicateReport>,
-    pub complexity: Option<ComplexityReport>,
-    pub api_surface: Option<ApiSurfaceReport>,
-    pub effort: Option<EffortEstimateReport>,
-    pub fun: Option<FunReport>,
-}
 
 // =========================
 // Ecosystem Envelope (v1) — re-exported from tokmd-envelope
